backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
numpy
tensorflow-lite
scikit-learn
python-multipart